use correct typecasting throughout

+++ src/t8_default/t8_default_hex.c
@@ -131,7 +131,7 @@ static void
 t8_default_hex_set_linear_id (t8_element_t * elem, int level, uint64_t id)
 {
   T8_ASSERT (0 <= level && level <= P8EST_QMAXLEVEL);
-  T8_ASSERT (0 <= id && id < (uint64_t) 1 << P8EST_DIM * level);
+  T8_ASSERT (0 <= id && id < ((uint64_t) 1) << P8EST_DIM * level);
 
   p8est_quadrant_set_morton ((p8est_quadrant_t *) elem, level, id);
 }
@@ -169,7 +169,7 @@ t8_default_hex_successor (const t8_element_t * elem1,
   T8_ASSERT (0 <= level && level <= P8EST_QMAXLEVEL);
 
   id = p8est_quadrant_linear_id ((const p8est_quadrant_t *) elem1, level);
-  T8_ASSERT (id + 1 < (1UL << P8EST_DIM * level));
+  T8_ASSERT (id + 1 < ((uint64_t) 1) << P8EST_DIM * level);
   p8est_quadrant_set_morton ((p8est_quadrant_t *) elem2, level, id + 1);
 }
