+++ src/t8_default/t8_default_quad.c
@@ -179,7 +179,7 @@ static void
 t8_default_quad_set_linear_id (t8_element_t * elem, int level, uint64_t id)
 {
   T8_ASSERT (0 <= level && level <= P4EST_QMAXLEVEL);
-  T8_ASSERT (0 <= id && id < (uint64_t) 1 << P4EST_DIM * level);
+  T8_ASSERT (0 <= id && id < ((uint64_t) 1) << P4EST_DIM * level);
 
   p4est_quadrant_set_morton ((p4est_quadrant_t *) elem, level, id);
   T8_QUAD_SET_TDIM ((p4est_quadrant_t *) elem, 2);
@@ -218,7 +218,7 @@ t8_default_quad_successor (const t8_element_t * elem1,
   T8_ASSERT (0 <= level && level <= P4EST_QMAXLEVEL);
 
   id = p4est_quadrant_linear_id ((const p4est_quadrant_t *) elem1, level);
-  T8_ASSERT (id + 1 < (1UL << P4EST_DIM * level));
+  T8_ASSERT (id + 1 < ((uint64_t) 1) << P4EST_DIM * level);
   p4est_quadrant_set_morton ((p4est_quadrant_t *) elem2, level, id + 1);
   t8_default_quad_copy_surround ((const p4est_quadrant_t *) elem1,
                                  (p4est_quadrant_t *) elem2);
